+++ src/app/components/rutinas/rutinas.component.ts
@@ -5,6 +5,7 @@ import { MatTableDataSource } from '@angular/material/table';
 import { ApiService } from 'src/app/services/api.service';
 import { FormRutinasComponent } from '../forms/form-rutinas/form-rutinas.component';
 import { MatDialog } from '@angular/material/dialog';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-rutinas',
@@ -13,17 +14,17 @@ import { MatDialog } from '@angular/material/dialog';
 })
 export class RutinasComponent implements OnInit, AfterViewInit {
 
-  displayedColumns: string[] = ['NombreRutina', 'CaloriasRutina', 'DescripcionRutina', 'CategoriaRutina', 'TiempoRutinaMin', 'acciones'];
+  displayedColumns: string[] = ['nombreRutina', 'caloriasRutina', 'descripcionRutina', 'categoriaRutina', 'tiempoRutinaMin', 'acciones'];
   dataSource: MatTableDataSource<any>;
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
 
   columnHeaders = {
-    NombreRutina: 'Nombre',
-    CaloriasRutina: 'Calorias',
-    DescripcionRutina: 'Descripcion',
-    CategoriaRutina: 'Categoria',
-    TiempoRutinaMin: 'Tiempo',
+    nombreRutina: 'Nombre',
+    caloriasRutina: 'Calorias',
+    descripcionRutina: 'Descripcion',
+    categoriaRutina: 'Categoria',
+    tiempoRutinaMin: 'Tiempo',
     acciones: 'Acciones',
   };
 
@@ -59,7 +60,24 @@ export class RutinasComponent implements OnInit, AfterViewInit {
   }
 
   removeRutina(rutina) {
-    this.apiService.delete('Rutinas', rutina.idRutina).then(res=>{this.ngOnInit()});
+    Swal.fire({
+      title: 'Esta Seguro que desea eliminar el registro?',
+      text: "Esta acción no se puede revertir",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.apiService.delete('Rutinas', rutina.idRutina).then(res=>{this.ngOnInit()});
+        Swal.fire(
+          'Registro Eliminado',
+          'El registro ha sido eliminado',
+          'success'
+        )
+      }
+    })
   }
 }
